# Replace pool with bytes in readLoop