Replace pool with bytes in readLoop

#### Description

#### Reference issue
Fixes #...