@Sean-Der 2024-04-24
@cnderrauber Would you mind explaining the reasoning for the change?

Does this reduce allocation for a single data channel? What is the impact if the user has many?

The original change was driven with data. Would be nice to match that!