@cnderrauber 2024-04-25
@Sean-Der It is not a revert to the change https://github.com/pion/webrtc/commit/159ba5aca3b96b85f7a653329102ea12046a1b1f but a simplify, the oldest behavior was allocating a 65KB buffer for each read call, then @bshi change it to get buffer from pool for every loop, but we don't need a pool here if the readLoop always need a bytes buffer, just allocate it once at the head and release it at end