@cnderrauber 2024-04-25
```
goos: darwin
goarch: arm64
pkg: github.com/pion/webrtc/v4
                     │    old.txt    │                  new.txt                  │
                     │    sec/op     │     sec/op      vs base                   │
DataChannelSend2-10     1.860µ ± ∞ ¹   1.837µ ±   ∞ ¹        ~ (p=1.000 n=1+3) ²
DataChannelSend4-10     3.882µ ± ∞ ¹   2.175µ ±   ∞ ¹        ~ (p=0.400 n=3+2) ²
DataChannelSend8-10     2.797µ ± ∞ ¹   3.271µ ±   ∞ ¹        ~ (p=1.000 n=1)   ²
DataChannelSend16-10   13.133µ ± ∞ ¹   5.842µ ±   ∞ ¹        ~ (p=0.333 n=2)   ²
DataChannelSend32-10     1.119 ± ∞ ¹    1.030 ± 19%          ~ (p=0.683 n=4+8)
geomean                 49.49µ         37.94µ          -23.33%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                     │    old.txt    │                  new.txt                  │
                     │     B/op      │      B/op        vs base                  │
DataChannelSend2-10    1.274Ki ± ∞ ¹   1.243Ki ±   ∞ ¹  -2.45% (n=1+3)
DataChannelSend4-10    1.305Ki ± ∞ ¹   1.252Ki ±   ∞ ¹       ~ (p=0.200 n=3+2) ²
DataChannelSend8-10    1.284Ki ± ∞ ¹   1.256Ki ±   ∞ ¹       ~ (p=1.000 n=1)   ²
DataChannelSend16-10   1.303Ki ± ∞ ¹   1.318Ki ±   ∞ ¹       ~ (p=1.000 n=2)   ²
DataChannelSend32-10   6.831Mi ± ∞ ¹   7.465Mi ± 22%         ~ (p=0.683 n=4+8)
geomean                7.208Ki         7.226Ki          +0.25%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                     │   old.txt    │                 new.txt                  │
                     │  allocs/op   │   allocs/op     vs base                  │
DataChannelSend2-10     33.00 ± ∞ ¹    32.00 ±   ∞ ¹       ~ (p=1.500 n=1+3) ²
DataChannelSend4-10     33.00 ± ∞ ¹    32.00 ±   ∞ ¹  -3.03% (n=3+2)
DataChannelSend8-10     33.00 ± ∞ ¹    32.00 ±   ∞ ¹       ~ (p=1.000 n=1)   ²
DataChannelSend16-10    24.50 ± ∞ ¹    32.00 ±   ∞ ¹       ~ (p=0.333 n=2)   ²
DataChannelSend32-10   87.43k ± ∞ ¹   73.69k ± 68%         ~ (p=0.808 n=4+8)
geomean                 150.4          150.5          +0.08%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```