@bshi 2024-04-26
> but we don't need a pool here if the readLoop always need a bytes buffer, just allocate it once at the head and release it at end

The reason that the pool is required is that, under high concurrency and GC pressure, memory allocation explodes and performance is crippled.

From the https://github.com/pion/webrtc/commit/159ba5aca3b96b85f7a653329102ea12046a1b1f - note how **allocations/op** stays constant as parallelism increases (and goes to 0 with a pooled approach).  Note also how time/op stays relatively constant because GC pressure has been alleviated.


![image](https://github.com/pion/webrtc/assets/15159/c9c529a4-633a-4c6e-91ad-1c44fa7a1feb)


Two things to note about the the most recent benchmark numbers in the previous comment are;

* the # of measurements needs to be higher to have more confidence in them (`-count`)
* The more interesting thing is that the old measurements already show signs of GC pressure.  At some point, some other changes in and around this code (not just this one) have made performance regressions.  It's probably worth repeating the profiling exercise in #1516 at this point.

```
                     │    old.txt    │                  new.txt                  │
                     │    sec/op     │     sec/op      vs base                   │
DataChannelSend2-10     1.860µ ± ∞ ¹   1.837µ ±   ∞ ¹        ~ (p=1.000 n=1+3) ²
DataChannelSend4-10     3.882µ ± ∞ ¹   2.175µ ±   ∞ ¹        ~ (p=0.400 n=3+2) ²
DataChannelSend8-10     2.797µ ± ∞ ¹   3.271µ ±   ∞ ¹        ~ (p=1.000 n=1)   ²
DataChannelSend16-10   13.133µ ± ∞ ¹   5.842µ ±   ∞ ¹        ~ (p=0.333 n=2)   ²
DataChannelSend32-10     1.119 ± ∞ ¹    1.030 ± 19%          ~ (p=0.683 n=4+8) <-- 1 second per operation (!!!)
geomean                 49.49µ         37.94µ          -23.33%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

I would recommend rolling back this change and profiling to figure out where this regression came from (or perhaps just bisect and benchmark for a quick search).  Instead of eliminating pooling, I suspect the team needs to *increase* the use of pools.

CC @Sean-Der